+++ app/src/main/java/com/sedilant/yambol/data/team/TrainEntity.kt
@@ -15,5 +15,5 @@ data class TrainEntity(
     val date: Long,
     val time: Float,
     val concepts: List<String>,
-    val teamId: Int
+    val teamId: Long
 )
